Restructure ASAP for flexible cost estimation.
There are now multiple variants of AsapPass, which are all subclasses of
AsapPassBase. These only differ in the type of cost estimation that they
use. For cost estimation, there is a similar set of subclasses of
SanityCheckCost.

Add a new type of cost computation based on covered program counters.
Each check that is on a covered source line has cost 1, all others have
cost zero.